refactored code into smaller components
