+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import FlatListScreen from './screens/FlatListScreen';
+import HomeScreen from './screens/HomeScreen';
 import MovieScreen from './screens/MovieScreen';
 import ComingSoonScreen from './screens/ComingSoonScreen'
 
@@ -12,8 +12,8 @@ const Stack = createNativeStackNavigator();
 const App = () => {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="FlatList">
-        <Stack.Screen name="FlatList" component={FlatListScreen} options={{ title: 'Home' }}  />
+      <Stack.Navigator initialRouteName="Home">
+        <Stack.Screen name="Home" component={HomeScreen} />
         <Stack.Screen name="Movie" component={MovieScreen} options={({ route }) => ({ title: route.params.title })}/>
         <Stack.Screen name="ComingSoonScreen" component={ComingSoonScreen} options={({ route }) => ({ title: route.params.title })}/>
       </Stack.Navigator>
