+++ components/Carousels/carouselStyles.js
@@ -1,0 +1,22 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  carouselHeaders: {
+    alignItems: "center",
+    fontSize: 23,
+    fontWeight: "bold",
+    margin: 15,
+  },
+
+  carousel: {
+    height: 350,
+    marginHorizontal: 5,
+  },
+
+  castCarousel: {
+    height: 250,
+    marginHorizontal: 5,
+  },
+});
+
+export default styles;
