+++ components/Tiles/ActorTile.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { View, Image, Text } from "react-native";
+
+import styles from "./TileStyles";
+
+const ActorTile = ({ item }) => {
+  return (
+    <View style={styles.actorTile}>
+      <Image style={styles.actorImage} source={{ uri: item.image }} />
+      <Text style={styles.actorName}>{item.title}</Text>
+    </View>
+  );
+};
+
+export default ActorTile;
