+++ screens/ComingSoonScreen.js
@@ -1,13 +1,15 @@
 import React from "react";
 import { SafeAreaView } from "react-native";
 
-import styles from "../styles";
-import ComingSoonTile from "../components/ComingSoonTile";
+import styles from "./ScreenStyles";
+import ComingSoonMovieInfo from "../components/Info/ComingSoonMovieInfo";
+import StarsCarousel from "../components/Carousels/StarsCarousel";
 
 const ComingSoonScreen = ({ route }) => {
   return (
     <SafeAreaView style={styles.movieScreenRoot}>
-      <ComingSoonTile item={route.params} />
+      <ComingSoonMovieInfo item={route.params} />
+      <StarsCarousel />
     </SafeAreaView>
   );
 };
