+++ screens/MovieScreen.js
@@ -1,13 +1,15 @@
 import React from "react";
 import { SafeAreaView } from "react-native";
 
-import styles from "../styles";
-import MovieTile from "../components/MovieTile";
+import styles from "./ScreenStyles";
+import Top250MovieInfo from '../components/Info/Top250MovieInfo'
+import StarsCarousel from "../components/Carousels/StarsCarousel";
 
 const MovieScreen = ({ route }) => {
   return (
     <SafeAreaView style={styles.movieScreenRoot}>
-      <MovieTile item={route.params} />
+      <Top250MovieInfo item={route.params} />
+      <StarsCarousel /> 
     </SafeAreaView>
   );
 };
