+++ screens/ScreenStyles.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  root: {
+    flex: 1,
+    justifyContent: "flex-start",
+    alignItems: "center",
+  },
+
+  movieScreenRoot: {
+    flex: 1,
+    justifyContent: "flex-start",
+    alignItems: "center",
+  },
+});
+
+export default styles;
